api/pyproject.toml
[project]
name = "aana-api"
version = "0.1.0"
description = "AANA - AI-Powered Search Engine Backend API"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "async-lru==2.0.4",
    "attrs==24.3.0",
    "beautifulsoup4==4.12.3",
    "Brotli==1.1.0",
    "cachetools==5.5.0",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "defusedxml==0.7.1",
    "dnspython==2.7.0",
    "email_validator==2.2.0",
    "fastapi==0.112.2",
    "fastapi-cli==0.0.7",
    "frozenlist==1.5.0",
    "google-auth==2.37.0",
    "google-genai==0.3.0",
    "h11==0.14.0",
    "html2text==2024.2.26",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "Jinja2==3.1.5",
    "lxml==5.3.0",
    "markdown-it-py==3.0.0",
    "MarkupSafe==3.0.2",
    "mdurl==0.1.2",
    "multidict==6.1.0",
    "pillow==11.0.0",
    "propcache==0.2.1",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.1",
    "pydantic==2.10.4",
    "pydantic_core==2.27.2",
    "Pygments==2.18.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "PyYAML==6.0.2",
    "requests==2.32.3",
    "rich==13.9.4",
    "rich-toolkit==0.12.0",
    "rsa==4.9",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "starlette==0.38.6",
    "typer==0.15.1",
    "typing_extensions==4.12.2",
    "urllib3==2.3.0",
    "uvicorn==0.30.6",
    "watchfiles==1.0.3",
    "websockets==14.1",
    "yarl==1.18.3",
    "youtube-transcript-api==0.6.3"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
